update version 3.0.0.11

+++ Classes/MSFWalletViewModel.m
@@ -139,7 +139,7 @@ static NSString *const kApplicationWalletType = @"4";
 					self.loanRates = [NSString stringWithFormat:@"%.3f%%", model.feeRate * 100];
 					self.repayDate = model.latestDueDate;
                     self.repayAmounts = model.latestDueMoney;
-                    self.repayDates = [NSString stringWithFormat:@"剩余%ld天,可随时还款", model.remainingDate];
+                    self.repayDates = [NSString stringWithFormat:@"剩余%ld天,可随时还款", (long)model.remainingDate];
 				}];
 			} break;
 			default:
